Use alternative WriteInPublic messages API
When retrieving messages through the API endpoint that doesn't contain
the instance ID WriteInPublic looks for people that are in _any_
instance who have the requested UUID, rather than just a specific
instance. This causes an error because the code uses .get(), which
expects one record to be returned, but if a person appears with the same
UUID in more than one instance then multiple records will be returned.

After some digging I found this alternative API endpoint, which contains
the instance ID (not the slug, unfortunately) in the URL, and therefore
only filters by contacts in that one instance.

This doesn't seem like an ideal solution, since it's using a different
API format to the other methods, but it works, so it's good enough for
now.