Output flash messages for WriteInPublic message sending
If the message is sent successfully then the user will be redirected to
the individual message page, so we need to output the flash messages
there.

If there is an error while sending the message then the user gets sent
back to the start of the message sending process, so the flash message
needs to be shown then.

Note: currently if there is an error then the users message is lost,
which isn't ideal, but it's at least _slightly_ better to send them back
to the start of the process rather than showing them a 500 error.